seperate server.go in different files
